Layout improvements
-Resizable/responsive ui
