+++ InfoPanel/ViewModels/UpdatesViewModel.cs
@@ -1,0 +1,47 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+using InfoPanel.Models;
+using System.Reflection;
+
+namespace InfoPanel.ViewModels
+{
+    public class UpdatesViewModel: ObservableObject
+    {
+        public string Version { get; set; }
+
+        public VersionModel? VersionModel { get; set; }
+
+        private bool _updateCheckInProgress = false;
+
+        public bool UpdateCheckInProgress
+        {
+            get { return _updateCheckInProgress; }
+            set { SetProperty(ref _updateCheckInProgress, value); }
+        }
+
+        private bool _downloadInProgress = false;
+        public bool DownloadInProgress
+        {
+            get { return _downloadInProgress; }
+            set { SetProperty(ref _downloadInProgress, value); }
+        }
+
+        private double _downloadProgress = 0;
+
+        public double DownloadProgress
+        {
+            get { return _downloadProgress; }
+            set { SetProperty(ref _downloadProgress, value); }
+        }
+
+        private bool _updateAvailable = false;
+        public bool UpdateAvailable
+        {
+            get { return _updateAvailable; }
+            set { SetProperty(ref _updateAvailable, value); }
+        }
+
+        public UpdatesViewModel() {
+            Version = Assembly.GetExecutingAssembly().GetName().Version!.ToString(3);
+        }
+    }
+}
