+++ InfoPanel/Views/Components/DisplayItems.xaml
@@ -10,7 +10,7 @@
              xmlns:ui="http://schemas.lepo.co/wpfui/2022/xaml"
              Foreground="{DynamicResource TextFillColorSecondaryBrush}">
 
-    <Grid Margin="10,10,10,10">
+    <Grid Margin="10,10,10,0">
         <Grid.RowDefinitions>
             <RowDefinition Height="auto"/>
             <RowDefinition Height="*"/>
@@ -182,19 +182,15 @@
             </Expander>
         </Grid>-->
 
-        <Grid Grid.Row="4" Margin="0,10,0,10">
+        <Grid Grid.Row="4" Margin="0,10,0,10" HorizontalAlignment="Right">
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="auto"/>
                 <ColumnDefinition Width="auto"/>
             </Grid.ColumnDefinitions>
-            <ui:Button x:Name="ButtonSave" Grid.Column="0" HorizontalAlignment="Left" Content="Save" Appearance="Primary"
-                       ToolTip="Make your changes permanent" Click="ButtonSave_Click" Margin="0,0,0,0" Width="150"/>
+            <ui:Button x:Name="ButtonSave" Grid.Column="0" Content="Save" Appearance="Primary"
+                       ToolTip="Make your changes permanent" Click="ButtonSave_Click" Margin="0,0,0,0" Width="140"/>
             <ui:Button x:Name="ButtonReload" Grid.Column="1" Icon="ArrowUndo20" Appearance="Danger" Content=""
                      ToolTip="Discard all changes and reload from disk" Click="ButtonReload_Click" Margin="10,0,0,0"/>
         </Grid>
-
     </Grid>
-
-
-
 </UserControl>
