+++ InfoPanel/Views/Components/HWiNFO/HWiNFOSensors.xaml
@@ -11,7 +11,7 @@
              Foreground="{DynamicResource TextFillColorSecondaryBrush}"
              mc:Ignorable="d">
 
-    <Grid Margin="10,0,10,10">
+    <Grid Margin="0,0,0,0">
         <Grid.RowDefinitions>
             <RowDefinition Height="Auto"/>
             <RowDefinition Height="1*"/>
